# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "heart",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/greet": {
      "get": {
        "tags": [
          "HelloWorld"
        ],
        "operationId": "greet",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Name to greet",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Greetings",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/hello": {
      "get": {
        "tags": [
          "HelloWorld"
        ],
        "operationId": "hello",
        "responses": {
          "200": {
            "description": "Say hello",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/load_session": {
      "get": {
        "tags": [
          "Session"
        ],
        "operationId": "load_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        }
      }
    },
    "/api/login": {
      "put": {
        "tags": [
          "Authenticate"
        ],
        "operationId": "authenticate",
        "parameters": [
          {
            "name": "user_name",
            "in": "query",
            "description": "User name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "Password",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "device_info",
            "in": "query",
            "description": "Device info",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/remove_session": {
      "delete": {
        "tags": [
          "Session"
        ],
        "operationId": "remove_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session removed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/session/count": {
      "get": {
        "tags": [
          "Session"
        ],
        "operationId": "count",
        "parameters": [
          {
            "name": "table_name",
            "in": "query",
            "description": "table name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "worked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountResponse"
                },
                "example": {
                  "count": 0
                }
              }
            }
          }
        }
      }
    },
    "/api/session/delete_all": {
      "delete": {
        "tags": [
          "Session"
        ],
        "operationId": "delete_all",
        "parameters": [
          {
            "name": "table_name",
            "in": "query",
            "description": "Table name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "worked"
          }
        }
      }
    },
    "/api/session/delete_by_expiry": {
      "delete": {
        "tags": [
          "Session"
        ],
        "operationId": "delete_by_expiry",
        "parameters": [
          {
            "name": "table_name",
            "in": "query",
            "description": "table name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "worked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/session/delete_one_by_id": {
      "delete": {
        "tags": [
          "Session"
        ],
        "operationId": "delete_one_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "table_name",
            "in": "query",
            "description": "Table name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "worked"
          }
        }
      }
    },
    "/api/session/exists": {
      "get": {
        "tags": [
          "Session"
        ],
        "operationId": "exists",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "table_name",
            "in": "query",
            "description": "Table name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "worked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExistsResponse"
                },
                "example": {
                  "value": false
                }
              }
            }
          }
        }
      }
    },
    "/api/session/get_ids": {
      "get": {
        "tags": [
          "Session"
        ],
        "operationId": "get_ids",
        "parameters": [
          {
            "name": "table_name",
            "in": "query",
            "description": "Table name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "worked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/session/load": {
      "get": {
        "tags": [
          "Session"
        ],
        "operationId": "load",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "table_name",
            "in": "query",
            "description": "Table name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "worked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResponse"
                },
                "example": {
                  "value": ""
                }
              }
            }
          },
          "404": {
            "description": "worked"
          }
        }
      }
    },
    "/api/session/store": {
      "put": {
        "tags": [
          "Session"
        ],
        "operationId": "store",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "session",
            "in": "query",
            "description": "Session data",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expires",
            "in": "query",
            "description": "Expiration timestamp",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "table_name",
            "in": "query",
            "description": "Table name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "worked"
          }
        }
      }
    },
    "/api/test": {
      "put": {
        "tags": [
          "Test"
        ],
        "operationId": "test_put",
        "parameters": [
          {
            "name": "test",
            "in": "path",
            "description": "Test",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test operation successful",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/todo": {
      "put": {
        "tags": [
          "Todo"
        ],
        "operationId": "todo_put",
        "parameters": [
          {
            "name": "todo",
            "in": "path",
            "description": "Json<Todo>",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo item created successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/update_session": {
      "put": {
        "tags": [
          "Session"
        ],
        "operationId": "update_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionValue"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session updated"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthenticateApi": {
        "default": null,
        "nullable": true
      },
      "AuthenticationResult": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Some"
            ],
            "properties": {
              "Some": {
                "type": "string"
              }
            },
            "example": "session_id"
          },
          {
            "type": "string",
            "enum": [
              "Conflict"
            ]
          },
          {
            "type": "string",
            "enum": [
              "Forbidden"
            ]
          }
        ]
      },
      "CountResponse": {
        "type": "object",
        "required": [
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ExistsResponse": {
        "type": "object",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "type": "boolean"
          }
        }
      },
      "GreetParams": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "LoadResponse": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "LoadSessionParams": {
        "type": "object",
        "required": [
          "session_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          }
        }
      },
      "LoginParams": {
        "type": "object",
        "required": [
          "user_name",
          "password",
          "device_info"
        ],
        "properties": {
          "device_info": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "user_name": {
            "type": "string"
          }
        }
      },
      "RemoveSessionParams": {
        "type": "object",
        "required": [
          "session_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          }
        }
      },
      "Test": {
        "type": "object",
        "required": [
          "test"
        ],
        "properties": {
          "test": {
            "type": "string"
          }
        }
      },
      "Todo": {
        "type": "object",
        "required": [
          "test"
        ],
        "properties": {
          "test": {
            "type": "string"
          }
        }
      },
      "UpdateSessionParams": {
        "type": "object",
        "required": [
          "session_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          }
        }
      },
      "UpdateSessionValue": {
        "type": "object",
        "required": [
          "entries"
        ],
        "properties": {
          "entries": {
            "type": "string"
          },
          "expires": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "tags": [
    {
      "name": "HelloWorld",
      "description": "Hello world operations"
    },
    {
      "name": "Authenticate",
      "description": "Authenticate operations"
    },
    {
      "name": "Test",
      "description": "Test operations"
    },
    {
      "name": "Todo",
      "description": "Todo operations"
    },
    {
      "name": "Session",
      "description": "Session operations"
    }
  ]
}